convert cash from double to int

+++ src/main/java/IncomeTaxSquare.java
@@ -8,7 +8,7 @@ public class IncomeTaxSquare extends Square {
 
     public void landedOn(Player p){
         int w = p.getNetWorth();
-        p.reduceCash(min(Math.floor(0.1 * w), 200));
+        p.reduceCash((int) min(Math.floor(0.1 * w), 200));
         p.setLocation(this);
     }
 }
